create build.sh

+++ html_generator/generate.rb
@@ -45,7 +45,7 @@ module View
   end
 end
 
-OUTPUT_DIR = File.join(File.dirname(__FILE__), 'dist').freeze
+OUTPUT_DIR = File.join(File.dirname(__FILE__), ARGV[0]).freeze
 TEMPLATE_PATH = File.join(File.dirname(__FILE__), 'templates', 'index.haml').freeze
 ICON_DIR = File.join(File.dirname(__FILE__), 'icons').freeze
 ENTRY_COUNT = 50.freeze
@@ -61,7 +61,7 @@ entries = JSON.parse(json_string)['entries'].map do |e|
   )
 end.sort_by(&:published_at).reverse.first(ENTRY_COUNT)
 
-FileUtils.rm_r(OUTPUT_DIR)
+FileUtils.rm_r(OUTPUT_DIR) if File.exist?(OUTPUT_DIR)
 Dir.mkdir(OUTPUT_DIR)
 result = Haml::Engine.new(File.read(TEMPLATE_PATH)).render(Object.new, entries: entries)
 output_path = File.join(OUTPUT_DIR, 'index.html')
